添加一些creation 函数

+++ utils.go
@@ -24,11 +24,11 @@ func ProductIntSlice(slice []int) int {
 	return prod
 }
 
-func GetShapeNum(shape []int64) int64 {
+func GetShapeNum(shape []int) int {
 	if len(shape) == 0 {
 		panic(fmt.Errorf("shape为空，无法计算数据个数。"))
 	}
-	var prod int64 = 1
+	var prod int = 1
 	for _, v := range shape {
 		if v <= 0 {
 			panic(fmt.Errorf("shape值必须为正数"))
